+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.38.0] - 2023-06-16
+
+### Changed
+
+- Add extra information customer
+
 ## [3.37.0] - 2023-06-12
 
 ### Changed